Added external interrupts and cleaned up some serial code.
Menu buttons work fam

+++ ADC.h
@@ -11,7 +11,7 @@
 void ADC1_init(void);
 /* Functions to read ADC and return value */
 double read_ADC1 (void);
-uint32_t read_ADC1_raw (void);
+uint16_t read_ADC1_raw (void);
 double scale_ADC1(uint16_t adc1_raw);
 
 /* function to set ADC to cont */
